Temp commit in case of losing data. Encountered WindowFromPoint issue, see https://github.com/dotnet/wpf/issues/6766.

+++ src/QHackCLR/Common.h
@@ -80,7 +80,7 @@ namespace QHackCLR {
 				return Equals(dynamic_cast<IClrHandled^>(obj));
 			}
 			int GetHashCode() override {
-				return m_ClrHandle.ToUInt32();
+				return (int)m_ClrHandle.ToUInt64();
 			}
 			static bool operator ==(ClrEntity^ a, ClrEntity^ b) {
 				if (Object::ReferenceEquals(a, nullptr))
